Create site.yml
Here is a sample site.yml file for a multi-tier architecture:

This file is the top-level playbook for the multi-tier architecture, and it specifies that the web, app, and db roles should be applied to all hosts in the inventory. The become directive allows the playbook to run with escalated privileges, and the gather_facts directive enables Ansible to gather information about the hosts before executing the tasks.

It is assumed that the web, app, and db roles have already been defined in separate directories, each containing the necessary tasks, templates, and files to configure the respective tier of the architecture. When the playbook is executed, Ansible will apply each of the roles in sequence, configuring the web, application, and database servers as defined in the roles.