Sample
Here is a sample db.yml file for configuring the database tier of a multi-tier architecture:
This playbook is intended to be used to configure the database servers in the multi-tier architecture. The hosts directive specifies that the playbook should only be executed on the servers designated as db in the inventory. The tasks section contains a series of actions that will be performed on each database server, including installing the database software, starting and enabling the service, and configuring the database using a template. The handlers section contains a handler that can be triggered by the notify directive in the configure database task to restart the database service if necessary.