Create app.yml
Here is a sample app.yml file for configuring the application tier of a multi-tier architecture:
This playbook is intended to be used to configure the application servers in the multi-tier architecture. The hosts directive specifies that the playbook should only be executed on the servers designated as app in the inventory. The tasks section contains a series of actions that will be performed on each application server, including installing required packages, installing application dependencies, copying the application code, and configuring the application service using a template. The handlers section contains a handler that can be triggered by the notify directive in the configure application service task to restart the application service if necessary.