Sample
Here is an example of a sample ansible playbook web.yml for a multi-tier architecture:
In this example, the playbook web.yml is responsible for configuring the web tier of the multi-tier architecture. The playbook uses the host group web from the inventory and runs the common and web roles. Additionally, the playbook has tasks for installing the Apache2 web server and starting the Apache2 service on the remote host. The become: yes flag is used to run the tasks as a privileged user (e.g., root).